+++ packages/agent-config/src/index.ts
@@ -1,6 +1,7 @@
 /**
  * @public
  */
+export * from './agentContextUtils'
 export * from './dataSources'
 export * from './agentCreator'
 export * from './objectCreator'