+++ packages/credential-store/src/types/ICredentialStore.ts
@@ -1,9 +1,16 @@
-import { IAgentContext, IPluginMethodMap } from '@veramo/core'
-import { CredentialRole, DigitalCredential, UpdateCredentialStateArgs } from '@sphereon/ssi-sdk.data-store'
-import { FindDigitalCredentialArgs } from '@sphereon/ssi-sdk.data-store/dist/types/digitalCredential/IAbstractDigitalCredentialStore'
-import { NonPersistedDigitalCredential } from '@sphereon/ssi-sdk.data-store/dist/types/digitalCredential/digitalCredential'
-import { FindClaimsArgs } from './claims'
+import { contextHasPlugin } from '@sphereon/ssi-sdk.agent-config'
+import {
+  CredentialRole,
+  DigitalCredential,
+  FindDigitalCredentialArgs,
+  NonPersistedDigitalCredential,
+  UpdateCredentialStateArgs,
+} from '@sphereon/ssi-sdk.data-store'
 import { ICredential, IPresentation, IVerifiableCredential, OriginalVerifiableCredential, OriginalVerifiablePresentation } from '@sphereon/ssi-types'
+import { IAgentContext, IPluginMethodMap } from '@veramo/core'
+import { FindClaimsArgs } from './claims'
+
+export type { UpdateCredentialStateArgs } // TODO create a local copy?
 
 export interface ICredentialStore extends IPluginMethodMap {
   /**
@@ -38,8 +45,7 @@ export interface ICredentialStore extends IPluginMethodMap {
 
   /**
    * Find one credential by id or hash
-   * @param CredentialRole
-   * @param idOrHash
+   * @param args
    */
   crsGetUniqueCredentialByIdOrHash(args: GetCredentialsByIdOrHashArgs): Promise<OptionalUniqueDigitalCredential>
 
@@ -66,6 +72,15 @@ export interface ICredentialStore extends IPluginMethodMap {
    * @param args
    */
   crsDeleteCredentials(args: DeleteCredentialsArgs): Promise<number>
+}
+
+/**
+ *
+ * @param context
+ * @internal
+ */
+export function contextHasCredentialStore(context: IAgentContext<IPluginMethodMap>): context is IAgentContext<ICredentialStore> {
+  return contextHasPlugin(context, 'crsGetCredential')
 }
 
 export type GetCredentialArgs = {
@@ -99,8 +114,6 @@ export type AddDigitalCredential = Omit<
 export type AddCredentialArgs = {
   credential: AddDigitalCredential
 }
-
-export type { UpdateCredentialStateArgs } from '@sphereon/ssi-sdk.data-store' // TODO create a local copy?
 
 export interface UniqueDigitalCredential {
   hash: string