+++ packages/siopv2-oid4vp-op-auth/src/session/OpSession.ts
@@ -8,17 +8,11 @@ import {
   VerificationMode,
   VerifiedAuthorizationRequest,
 } from '@sphereon/did-auth-siop'
-import { getAgentDIDMethods, getAgentResolver, getDID } from '@sphereon/ssi-sdk-ext.did-utils'
+import { getAgentDIDMethods, getAgentResolver } from '@sphereon/ssi-sdk-ext.did-utils'
 import { CredentialMapper, parseDid } from '@sphereon/ssi-types'
 import { IIdentifier, TKeyType } from '@veramo/core'
 import Debug from 'debug'
-import {
-  IOPOptions,
-  IOpSessionArgs,
-  IOpsSendSiopAuthorizationResponseArgs,
-  IRequiredContext,
-  IOpSessionGetOID4VPArgs,
-} from '../types/IDidAuthSiopOpAuthenticator'
+import { IOPOptions, IOpSessionArgs, IOpSessionGetOID4VPArgs, IOpsSendSiopAuthorizationResponseArgs, IRequiredContext } from '../types'
 import { createOP } from './functions'
 import { OID4VP } from './OID4VP'
 
@@ -265,7 +259,7 @@ export class OpSession {
       context: this.context,
     })
 
-    let issuer = args.responseSignerOpts?.issuer ?? (args.responseSignerOpts?.identifier ? getDID(args.responseSignerOpts) : undefined)
+    let issuer = args.responseSignerOpts.issuer
     const responseOpts = {
       verification,
       issuer,