+++ packages/siopv2-oid4vp-op-auth/src/types/IDidAuthSiopOpAuthenticator.ts
@@ -12,7 +12,11 @@ import {
 } from '@sphereon/did-auth-siop'
 import { DIDDocument } from '@sphereon/did-uni-client'
 import { VerifiablePresentationResult } from '@sphereon/pex'
-import { IIdentifierOpts } from '@sphereon/ssi-sdk-ext.did-utils'
+import { IIdentifierResolution, ManagedIdentifierOpts } from '@sphereon/ssi-sdk-ext.identifier-resolution'
+import { ICredentialStore } from '@sphereon/ssi-sdk.credential-store'
+import { Party } from '@sphereon/ssi-sdk.data-store'
+import { IPDManager } from '@sphereon/ssi-sdk.pd-manager'
+import { ISDJwtPlugin } from '@sphereon/ssi-sdk.sd-jwt'
 import { Hasher, PresentationSubmission, W3CVerifiableCredential, W3CVerifiablePresentation } from '@sphereon/ssi-types'
 import { VerifyCallback } from '@sphereon/wellknown-dids-client'
 import {
@@ -27,10 +31,7 @@ import {
 } from '@veramo/core'
 import { EventEmitter } from 'events'
 import { OpSession } from '../session'
-import { IPDManager } from '@sphereon/ssi-sdk.pd-manager'
-import { ISDJwtPlugin } from '@sphereon/ssi-sdk.sd-jwt'
 import { Siopv2Machine as Siopv2MachineId } from './machine'
-import { Party } from '@sphereon/ssi-sdk.data-store'
 import {
   AddIdentityArgs,
   CreateConfigArgs,
@@ -45,33 +46,40 @@ import {
   Siopv2AuthorizationRequestData,
   Siopv2AuthorizationResponseData,
 } from './siop-service'
-import { ICredentialStore } from '@sphereon/ssi-sdk.credential-store'
 
 export const LOGGER_NAMESPACE = 'sphereon:siopv2-oid4vp:op-auth'
 
 export interface IDidAuthSiopOpAuthenticator extends IPluginMethodMap {
   siopGetOPSession(args: IGetSiopSessionArgs, context: IRequiredContext): Promise<OpSession>
+
   siopRegisterOPSession(args: Omit<IOpSessionArgs, 'context'>, context: IRequiredContext): Promise<OpSession>
+
   siopRemoveOPSession(args: IRemoveSiopSessionArgs, context: IRequiredContext): Promise<boolean>
+
   siopRegisterOPCustomApproval(args: IRegisterCustomApprovalForSiopArgs, context: IRequiredContext): Promise<void>
+
   siopRemoveOPCustomApproval(args: IRemoveCustomApprovalForSiopArgs, context: IRequiredContext): Promise<boolean>
 
   siopGetMachineInterpreter(args: GetMachineArgs, context: RequiredContext): Promise<Siopv2MachineId>
+
   siopCreateConfig(args: CreateConfigArgs): Promise<CreateConfigResult>
+
   siopGetSiopRequest(args: GetSiopRequestArgs, context: RequiredContext): Promise<Siopv2AuthorizationRequestData>
+
   siopRetrieveContact(args: RetrieveContactArgs, context: RequiredContext): Promise<Party | undefined>
+
   siopAddIdentity(args: AddIdentityArgs, context: RequiredContext): Promise<void>
+
   siopSendResponse(args: SendResponseArgs, context: RequiredContext): Promise<Siopv2AuthorizationResponseData>
+
   siopGetSelectableCredentials(args: GetSelectableCredentialsArgs, context: RequiredContext): Promise<SelectableCredentialsMap>
 }
 
 export interface IOpSessionArgs {
   sessionId?: string
-
   requestJwtOrUri: string | URI
   providedPresentationDefinitions?: Array<PresentationDefinitionWithLocation>
-  idOpts?: IIdentifierOpts
-  // identifier: IIdentifier
+  identifierOptions?: ManagedIdentifierOpts
   context: IRequiredContext
   op?: IOPOptions
 }
@@ -108,7 +116,7 @@ export interface IRemoveCustomApprovalForSiopArgs {
 }
 
 export interface IOpsSendSiopAuthorizationResponseArgs {
-  responseSignerOpts: IIdentifierOpts & { issuer?: string; kid?: string }
+  responseSignerOpts: ManagedIdentifierOpts
   // verifiedAuthorizationRequest: VerifiedAuthorizationRequest
   presentationSubmission?: PresentationSubmission
   verifiablePresentations?: W3CVerifiablePresentation[]
@@ -119,7 +127,16 @@ export enum events {
 }
 
 export type IRequiredContext = IAgentContext<
-  IDataStoreORM & IResolver & IDIDManager & IKeyManager & ICredentialIssuer & ICredentialVerifier & ICredentialStore & IPDManager & ISDJwtPlugin
+  IDataStoreORM &
+    IResolver &
+    IDIDManager &
+    IKeyManager &
+    IIdentifierResolution &
+    ICredentialIssuer &
+    ICredentialVerifier &
+    ICredentialStore &
+    IPDManager &
+    ISDJwtPlugin
 >
 
 export interface IOPOptions {
@@ -137,6 +154,7 @@ export interface IOPOptions {
 
   resolveOpts?: ResolveOpts
 }
+
 /*
 export interface IIdentifierOpts {
   identifier: IIdentifier
@@ -152,7 +170,7 @@ export interface VerifiableCredentialsWithDefinition {
 export interface VerifiablePresentationWithDefinition extends VerifiablePresentationResult {
   definition: PresentationDefinitionWithLocation
   verifiableCredentials: W3CVerifiableCredential[]
-  identifierOpts: IIdentifierOpts
+  idOpts: ManagedIdentifierOpts
 }
 
 export interface IOpSessionGetOID4VPArgs {