+++ packages/siopv2-oid4vp-op-auth/src/types/machine/index.ts
@@ -1,5 +1,5 @@
 import { VerifiedAuthorizationRequest } from '@sphereon/did-auth-siop'
-import { IIdentifierOpts } from '@sphereon/ssi-sdk-ext.did-utils'
+import { ManagedIdentifierOpts } from '@sphereon/ssi-sdk-ext.identifier-resolution'
 import { DidAuthConfig, Party } from '@sphereon/ssi-sdk.data-store'
 import { OriginalVerifiableCredential } from '@sphereon/ssi-types'
 import { BaseActionObject, Interpreter, ResolveTypegenMeta, ServiceMap, State, StateMachine, TypegenDisabled } from 'xstate'
@@ -8,7 +8,7 @@ import { SelectableCredentialsMap, Siopv2AuthorizationRequestData, Siopv2Authori
 
 export type Siopv2MachineContext = {
   url: string
-  idOpts?: IIdentifierOpts
+  idOpts?: ManagedIdentifierOpts
   didAuthConfig?: Omit<DidAuthConfig, 'identifier'>
   authorizationRequestData?: Siopv2AuthorizationRequestData
   authorizationResponseData?: Siopv2AuthorizationResponseData
@@ -75,7 +75,7 @@ export type Siopv2StateMachine = StateMachine<
 
 export type CreateSiopv2MachineOpts = {
   url: string | URL
-  idOpts?: IIdentifierOpts
+  idOpts?: ManagedIdentifierOpts
   machineId?: string
 }